+++ src/AppBundle/Entity/User.php
@@ -3,6 +3,7 @@
 namespace AppBundle\Entity;
 
 use AppBundle\Entity\Traits\LoginInfoTrait;
+use AppBundle\Entity\Traits\AddressTrait;
 use JMS\Serializer\Annotation as JMS;
 use Symfony\Component\Security\Core\User\AdvancedUserInterface;
 use Symfony\Component\Validator\Constraints as Assert;
@@ -14,6 +15,7 @@ use AppBundle\Validator\Constraints\EmailSameDomain;
 class User implements AdvancedUserInterface
 {
     use LoginInfoTrait;
+    use AddressTrait;
 
     const ROLE_ADMIN = 'ROLE_ADMIN';
     const ROLE_LAY_DEPUTY = 'ROLE_LAY_DEPUTY';
@@ -157,53 +159,6 @@ class User implements AdvancedUserInterface
      * @var string
      */
     private $gaTrackingId;
-
-    /**
-     * @JMS\Type("string")
-     * @JMS\Groups({"user_details_full"})
-     * @Assert\NotBlank( message="user.address1.notBlank", groups={"user_details_full"} )
-     * @Assert\Length( max=200, maxMessage="user.address1.maxMessage", groups={"user_details_full"} )
-     *
-     * @var string
-     */
-    private $address1;
-
-    /**
-     * @JMS\Type("string")
-     * @JMS\Groups({"user_details_full"})
-     * @Assert\Length( max=200, maxMessage="user.address1.maxMessage", groups={"user_details_full"} )
-     *
-     * @var string
-     */
-    private $address2;
-
-    /**
-     * @JMS\Type("string")
-     * @JMS\Groups({"user_details_full"})
-     * @Assert\Length( max=200, maxMessage="user.address1.maxMessage", groups={"user_details_full"} )
-     *
-     * @var string
-     */
-    private $address3;
-
-    /**
-     * @JMS\Type("string")
-     * @JMS\Groups({"user_details_full"})
-     * @Assert\NotBlank( message="user.addressPostcode.notBlank", groups={"user_details_full"} )
-     * @Assert\Length(min=2, max=10, minMessage="user.addressPostcode.minLength", maxMessage="user.addressPostcode.maxLength", groups={"user_details_full"} )
-     *
-     * @var string
-     */
-    private $addressPostcode;
-
-    /**
-     * @JMS\Type("string")
-     * @JMS\Groups({"user_details_full"})
-     * @Assert\NotBlank( message="user.addressCountry.notBlank", groups={"user_details_full"} )
-     *
-     * @var string
-     */
-    private $addressCountry;
 
     /**
      * @JMS\Type("string")
@@ -594,31 +549,6 @@ class User implements AdvancedUserInterface
         return [$this->roleName];
     }
 
-    public function getAddress1()
-    {
-        return $this->address1;
-    }
-
-    public function getAddress2()
-    {
-        return $this->address2;
-    }
-
-    public function getAddress3()
-    {
-        return $this->address3;
-    }
-
-    public function getAddressPostcode()
-    {
-        return $this->addressPostcode;
-    }
-
-    public function getAddressCountry()
-    {
-        return $this->addressCountry;
-    }
-
     public function getPhoneMain()
     {
         return $this->phoneMain;
@@ -627,31 +557,6 @@ class User implements AdvancedUserInterface
     public function getPhoneAlternative()
     {
         return $this->phoneAlternative;
-    }
-
-    public function setAddress1($address1)
-    {
-        $this->address1 = $address1;
-    }
-
-    public function setAddress2($address2)
-    {
-        $this->address2 = $address2;
-    }
-
-    public function setAddress3($address3)
-    {
-        $this->address3 = $address3;
-    }
-
-    public function setAddressPostcode($addressPostcode)
-    {
-        $this->addressPostcode = $addressPostcode;
-    }
-
-    public function setAddressCountry($addressCountry)
-    {
-        $this->addressCountry = $addressCountry;
     }
 
     public function setPhoneMain($phoneMain)
